refactor: move demographic questionnaire from final to debrief
